Add files via upload
This is an idealized model of people waiting for bus coming, with a distance of approximately three kilometers and five traffic lights. The program creates a text file on your Desktop to show data collected from the simulation. You could tell the program how many times you want to run the simulation, and the numbers in the list returned indicates the total time a person needs to wait in seconds (if there are nothing in the list, it means that there are nobody waiting for the bus in this time of simulation, and each element in the list indicates one person's waiting time). The data are sorted and printed out.